refactor(#5): split workflow settings tab into cards and hooks
Extract static/MEDDPICC cards, freigabe settings, simulation panel, and settings/simulation hooks so WorkflowTab stays a thin composer.
